extract outlineitemwdg
